% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PCAseq.R
\docType{package}
\name{PCAseq}
\alias{PCAseq}
\alias{PCAseq-package}
\title{PCAseq: Genetic Population Sturcture Inference with Sequence Data.}
\description{
This package builds on the SNPRelate package by implmeting the PCA-seq method
of population inference for sequence data in a way that is compatible with
existing functions in SNPRelate.
}
\details{
The only function you're likely to need from \pkg{roxygen2} is
\code{\link{roxygenize}}. Otherwise refer to the vignettes to see
how to format the documentation.
}

