\name{PCAseq-package}
\alias{PCAseq-package}
\alias{PCAseq}
\docType{package}
\title{
PCAseq: Population Inference with Sequence Data}
\description{
This package builds on the SNPRelate package by implmeting the PCA-seq method of population inference with sequence data in a way that is compatible with existing functions in SNPRelate.
}
\details{
\tabular{ll}{
Package: \tab PCA-seq\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-16\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Jennifer L Kirk

Maintainer: Jennifer L Kirk <jlkirk@uw.edu>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
