% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqPCA.R
\name{seqPCA}
\alias{seqGRM}
\alias{seqPCA}
\title{Run PCA-seq on a GDS file}
\usage{
seqPCA(gdsobj, weights = c(1, 1), sample.id = NULL, snp.id = NULL,
  autosome.only = TRUE, remove.monosnp = TRUE, maf = NA,
  missing.rate = NaN, eigen.cnt = 32, need.genmat = FALSE,
  verbose = TRUE)
}
\arguments{
\item{gdsobj}{an object of the class SNPGDSFileClass, a SNP GDS file.}

\item{weights}{a vector of two numbers, indicating the paraters to use 
for the beta function weights; see Details.}

\item{sample.id}{a vector of sample ids specifying the samples to use for
analysis; if NULL, all samples are used.}

\item{snp.id}{a vector of SNP ids specifying the SNPs to use for analysis;
if NULL, all SNPs are used.}

\item{autosome.only}{if TRUE, use autosomal SNPs only; if it is a numeric or
character vector, keep SNPs according to the specified chromosomes.}

\item{remove.monosnp}{if TURE, remove monomorphic SNPs.}

\item{maf}{a string of the form "<min, max>" where "<" may be "(" or "[" and
">" may be ")" or "]"; this indicates the minimum and maximum MAF to allow}

\item{missing.rate}{to use the SNPs with missing rates less than or equal to
missing.rate; if NaN, no misisng threshold.}

\item{eigen.cnt}{the number of eigen vectors and values to return; if zero,
return all eigenvalues and vectors.}

\item{need.genmat}{if TRUE, return the genetic relatedness matrix.}

\item{verbose}{Not supported.}
}
\value{
Return a \code{snpPCAClass} object, a list with the follow slots:
\describe{
\item{\code{weights}}{the parameters used to define the weights
 used to calculate the GRM}
\item{\code{maf}}{the MAF cutoffs used}
\item{\code{sample.id}}{the sample ids used in the analysis}
\item{\code{snp.id}}{the SNP ids used in the analysis}
\item{\code{eigenval}}{eigenvalues}
\item{\code{eigenvect}}{a matrix of eigenvectors of dimensions # of samples by
\code{eigen.cnt}}
\item{\code{varprop}}{the proportion of the variance explained by each principal
component}
\item{\code{TraceXTX}}{the trace of the genetic relateness matrix}
\item{\code{Bayesian}}{indicates Bayes normalization; set to FALSE, as this is not
currently supported}
\item{\code{genmat}}{the genetic relateness matrix}}
}
\description{
This function calculates the Genetic Relatedness Matrix (GRM) on a GDS file
using the PCA-seq method for sequence data.
}
\details{
The weights parameter is a vector of two numbers, (a, b), that is 
passed to a weighting function based on a Beta(a,b). This function uses the
density of this Beta function at the SNP's minor allele frequency as the
weight for the SNP.
}
\section{Note}{

If you need to run the EIGENSTRAT method on a very large data set and do not
need to subset by both a minimum and maximum MAF, the
\code{\link[SNPRelate]{snpgdsPCA}} function will be faster.
}

